+++ lib/CallableExpectedException.php
@@ -20,28 +20,7 @@
 
 namespace Opis\Routing;
 
-class Pattern
+class CallableExpectedException extends \InvalidArgumentException
 {
-    protected $pattern;
-    
-    public function __construct($pattern)
-    {
-        $this->pattern = $pattern;
-    }
-    
-    public function serialize()
-    {
-        return serialize($this->pattern);
-    }
-    
-    public function unserialize($data)
-    {
-        $this->pattern = unserialize($data);
-    }
-    
-    public function __toString()
-    {
-        return $this->pattern;
-    }
     
 }