+++ lib/Collections/FilterCollection.php
@@ -21,7 +21,7 @@
 namespace Opis\Routing\Collections;
 
 use InvalidArgumentException;
-use Opis\Routing\Contracts\FilterInterface;
+use Opis\Routing\FilterInterface;
 
 class FilterCollection extends AbstractCollection
 {
@@ -30,7 +30,7 @@ class FilterCollection extends AbstractCollection
     {
         if(!($value instanceof FilterInterface))
         {
-            throw new InvalidArgumentException('Expected \Opis\Routing\Contracts\FilterInterface');
+            throw new InvalidArgumentException('Expected \Opis\Routing\FilterInterface');
         }
     }