+++ lib/DispatcherResolver.php
@@ -20,13 +20,9 @@
 
 namespace Opis\Routing;
 
-use Opis\Routing\Contracts\DispatcherResolverInterface;
-use Opis\Routing\Contracts\PathInterface;
-use Opis\Routing\Contracts\RouteInterface;
-
-class DispatcherResolver implements DispatcherResolverInterface
+class DispatcherResolver
 {
-    public function resolve(PathInterface $path, RouteInterface $route)
+    public function resolve(Path $path, Route $route)
     {
         return new Dispatcher();
     }